Fix error estimate on the strain-stress fit
As of 0.7 branch, scipy aparantly reports the standard
estimate of the error correctly so we should not need to
re-correct. This correction is thus made conditional on
the scipy version.

However, with my version of scipy, I get clearly wrong
errors (confirmed by fitting in gnuplot) so calculate
the error on the grad (i.e. the elastic constant) 'by
hand' and don't use scipy's stderr return value. This
approach is used for version 0.7 and above.