Propogate fitting errors into Cij
Propogate the errors in the fits into the Cij matix
assuming that fits are not correlated with each other.
Report the errors as a matrix. We can use this to calculate
errors on Sij or the various other bits and peices (and rotate the
error matrix if we need to).