Update regexp for upcoming CASTEP 6.1 release
Dominik Jochym reports that spaces will be
added before colons in a couple of places in the
castep output file. This breaks the existing
regular expressions used to find the final cell
parameters to set up the strained system. This
patch, from Dominik at STFC, allows the spaces
but maintains the ability to use older (pre-6.1)
files.