Fix error calculation
New (> 0.10.0) versions of scipy were incorrectly changing the calculation
of the error on the elastic constants due to a change in scipy.stats.linres
before version 0.7.0. This corrects the test and directly uses the standard
error from the fit.

Thanks to Keith Refson for reporting this and for providing the fix.