Initia pass at reading data from a cell file
This needs testing and a review, but it looks like it works with,
e.g.

    ../../generate_strain.py --use-cell MgO_primitive.cell -l 5 MgO_primitive

NB: won't work for all pointgroups (see FIXME in generate_strain), needs
to check it works (see FIXME in castep.py) and could use more flexibility
in the kind of cell file it reads (lattice or abc, frac or cart).